+++ tests/src/rules/no-unused-modules.js
@@ -1,6 +1,8 @@
 import { test, testFilePath, getTSParsers } from '../utils'
 import jsxConfig from '../../../config/react'
 import typescriptConfig from '../../../config/typescript'
+import eslintPkg from 'eslint/package.json'
+import semver from 'semver'
 
 import { RuleTester } from 'eslint'
 import fs from 'fs'
@@ -103,32 +105,39 @@ ruleTester.run('no-unused-modules', rule, {
 // tests for  exports
 ruleTester.run('no-unused-modules', rule, {
   valid: [
-
     test({ options: unusedExportsOptions,
            code: 'import { o2 } from "./file-o";export default () => 12',
-           filename: testFilePath('./no-unused-modules/file-a.js')}),
+           filename: testFilePath('./no-unused-modules/file-a.js'),
+           parser: require.resolve('babel-eslint')}),
     test({ options: unusedExportsOptions,
            code: 'export const b = 2',
-           filename: testFilePath('./no-unused-modules/file-b.js')}),
+           filename: testFilePath('./no-unused-modules/file-b.js'),
+           parser: require.resolve('babel-eslint')}),
     test({ options: unusedExportsOptions,
            code: 'const c1 = 3; function c2() { return 3 }; export { c1, c2 }',
-           filename: testFilePath('./no-unused-modules/file-c.js')}),
+           filename: testFilePath('./no-unused-modules/file-c.js'),
+           parser: require.resolve('babel-eslint')}),
     test({ options: unusedExportsOptions,
            code: 'export function d() { return 4 }',
-           filename: testFilePath('./no-unused-modules/file-d.js')}),
+           filename: testFilePath('./no-unused-modules/file-d.js'),
+           parser: require.resolve('babel-eslint')}),
     test({ options: unusedExportsOptions,
            code: 'export class q { q0() {} }',
-           filename: testFilePath('./no-unused-modules/file-q.js')}),
+           filename: testFilePath('./no-unused-modules/file-q.js'),
+           parser: require.resolve('babel-eslint')}),
     test({ options: unusedExportsOptions,
            code: 'const e0 = 5; export { e0 as e }',
-           filename: testFilePath('./no-unused-modules/file-e.js')}),
+           filename: testFilePath('./no-unused-modules/file-e.js'),
+           parser: require.resolve('babel-eslint')}),
     test({ options: unusedExportsOptions,
            code: 'const l0 = 5; const l = 10; export { l0 as l1, l }; export default () => {}',
-           filename: testFilePath('./no-unused-modules/file-l.js')}),
+           filename: testFilePath('./no-unused-modules/file-l.js'),
+           parser: require.resolve('babel-eslint')}),
     test({ options: unusedExportsOptions,
            code: 'const o0 = 0; const o1 = 1; export { o0, o1 as o2 }; export default () => {}',
-           filename: testFilePath('./no-unused-modules/file-o.js')}),
-    ],
+           filename: testFilePath('./no-unused-modules/file-o.js'),
+           parser: require.resolve('babel-eslint')}),
+  ],
   invalid: [
     test({ options: unusedExportsOptions,
            code: `import eslint from 'eslint'
@@ -194,6 +203,64 @@ ruleTester.run('no-unused-modules', rule, {
            filename: testFilePath('./no-unused-modules/file-k.js'),
            errors: [error(`exported declaration 'k' not used within other modules`)]}),
   ],
+})
+
+
+describe('dynamic imports', () => { 
+  if (semver.satisfies(eslintPkg.version, '< 6')) {
+    this.skip()
+    return
+  }
+
+  // test for unused exports with `import()`
+  ruleTester.run('no-unused-modules', rule, {
+    valid: [
+      test({ options: unusedExportsOptions,
+            code: `
+            export const a = 10
+            export const b = 20
+            export const c = 30
+            const d = 40
+            export default d
+            `,
+            parser: require.resolve('babel-eslint'),
+            filename: testFilePath('./no-unused-modules/exports-for-dynamic-js.js')}),
+    ],
+    invalid: [
+      test({ options: unusedExportsOptions,
+        code: `
+        export const a = 10
+        export const b = 20
+        export const c = 30
+        const d = 40
+        export default d
+        `,
+        parser: require.resolve('babel-eslint'),
+        filename: testFilePath('./no-unused-modules/exports-for-dynamic-js-2.js'),
+        errors: [
+            error(`exported declaration 'a' not used within other modules`),
+            error(`exported declaration 'b' not used within other modules`),
+            error(`exported declaration 'c' not used within other modules`),
+            error(`exported declaration 'default' not used within other modules`),
+        ]}),
+    ],
+  })
+  typescriptRuleTester.run('no-unused-modules', rule, {
+    valid: [
+      test({ options: unusedExportsTypescriptOptions,
+            code: `
+            export const ts_a = 10
+            export const ts_b = 20
+            export const ts_c = 30
+            const ts_d = 40
+            export default ts_d
+            `,
+            parser: require.resolve('@typescript-eslint/parser'),
+            filename: testFilePath('./no-unused-modules/typescript/exports-for-dynamic-ts.ts')}),
+    ],
+    invalid: [
+    ],
+  })
 })
 
 // // test for export from